fixed all register controller bugs

+++ backend/src/utils/cloudinary.js
@@ -3,7 +3,7 @@ import fs from "fs"    //File system of node
 
 cloudinary.config({
     cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
-    api_key: process.envCLOUDINARY_KEY,
+    api_key: process.env.CLOUDINARY_KEY,
     api_secret: process.env.CLOUDINARY_SECRET
 });
 
@@ -14,10 +14,14 @@ const uploadCloudinary = async (localFilePath)=>{
         const res = await cloudinary.uploader.upload(localFilePath, {
             resource_type: "auto"
         })
-        console.log("File is uploaded on cloudinary", res.url);
+        //console.log("File is uploaded on cloudinary", res.url);
+        fs.unlinkSync(localFilePath);
         return res;
     }catch(error){
-        fs.unlinkSync(localFilePath)   //Delete the locally saved file as the upload operation got faile
+        console.error("Cloudinary Upload Error: ", error);
+        if(fs.existsSync(localFilePath)){
+            fs.unlinkSync(localFilePath)   //Delete the locally saved file as the upload operation got faile
+        }
         return null;
     }
 }
